refactor(monitoring): extract tests via #[path] (cycle 40, 302→252 LOC)
